+++ source/src/nemo-application.c
@@ -89,6 +89,7 @@
 #include <libnemo-private/nemo-appearance.h>
 
 #include <libnemo-private/nemo-desktop-thumbnail.h>
+#include <libnemo-private/nemo-thumbnail-prune.h>
 
 #define NEMO_ACCEL_MAP_SAVE_DELAY 30
 
@@ -651,6 +652,8 @@ nemo_application_startup (GApplication *app)
 
     self->priv->cache_problem = FALSE;
     self->priv->ignore_cache_problem = FALSE;
+
+    nemo_thumbnail_prune_schedule ();
 
     /* If 'treat-root-as-normal' is true, assume we're running root as well,
        so we can skip the permission checks */